Optimize Known NPCs finding
This is the most fast, flexible and reliable implementation I may come up with.

It gets the data directly from the ActorBase in Skyrim rather than filtering the database in hopes to find it.

This in turn accelerates generic NPC data finding, since now it doesn't have to prove it's not a Known NPC.