Add archetype finding by name
Archetype solving will use name or class.
This was done to cover all those cases when an NPC named "soldier" has a "dremora" class.

Ignore the "EasyContainers" stuff.
That's a quick test I made for another mod, but since all the foundation to make it work was already done here, I just put it in here.

Will eventually move it to a proper mod.